Move resolver code from library to main binary
